+++ Common/src/main/java/de/markusbordihn/easymodelentities/item/EasyModelSpawnItem.java
@@ -1,0 +1,58 @@
+/*
+ * Copyright 2026 Markus Bordihn
+ *
+ * Permission is hereby granted, free of charge, to any person obtaining a copy of this software and
+ * associated documentation files (the "Software"), to deal in the Software without restriction,
+ * including without limitation the rights to use, copy, modify, merge, publish, distribute,
+ * sublicense, and/or sell copies of the Software, and to permit persons to whom the Software is
+ * furnished to do so, subject to the following conditions:
+ *
+ * The above copyright notice and this permission notice shall be included in all copies or
+ * substantial portions of the Software.
+ *
+ * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR IMPLIED, INCLUDING BUT
+ * NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY, FITNESS FOR A PARTICULAR PURPOSE AND
+ * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM,
+ * DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
+ * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
+ */
+
+package de.markusbordihn.easymodelentities.item;
+
+import java.util.function.Consumer;
+import net.minecraft.network.chat.Component;
+import net.minecraft.world.entity.player.Player;
+import net.minecraft.world.item.Item;
+import net.minecraft.world.item.ItemStack;
+import net.minecraft.world.item.TooltipFlag;
+import net.minecraft.world.item.component.TooltipDisplay;
+
+public abstract class EasyModelSpawnItem extends Item {
+
+  protected EasyModelSpawnItem(Properties properties) {
+    super(properties);
+  }
+
+  protected static void notifyPlayer(Player player, String message) {
+    if (player != null) {
+      player.sendSystemMessage(Component.literal(message));
+    }
+  }
+
+  @Override
+  public Component getName(ItemStack stack) {
+    return EasyModelEntitiesItems.profileId(stack)
+        .<Component>map(id -> Component.literal(EasyModelEntitiesItems.displayName(id)))
+        .orElseGet(() -> super.getName(stack));
+  }
+
+  @Override
+  public void appendHoverText(
+      ItemStack stack,
+      Item.TooltipContext context,
+      TooltipDisplay tooltipDisplay,
+      Consumer<Component> tooltipAdder,
+      TooltipFlag flag) {
+    EasyModelItemTooltip.append(stack, tooltipAdder);
+  }
+}