+++ Forge/src/main/java/de/markusbordihn/easymodelentities/EasyModelBlockEntityTypes.java
@@ -36,10 +36,9 @@ import net.minecraftforge.registries.DeferredRegister;
 import net.minecraftforge.registries.ForgeRegistries;
 import net.minecraftforge.registries.RegistryObject;
 
-public final class ForgeEasyModelBlockEntityTypes implements EasyModelHostBlockEntityTypeProvider {
+public final class EasyModelBlockEntityTypes implements EasyModelHostBlockEntityTypeProvider {
 
-  public static final ForgeEasyModelBlockEntityTypes INSTANCE =
-      new ForgeEasyModelBlockEntityTypes();
+  public static final EasyModelBlockEntityTypes INSTANCE = new EasyModelBlockEntityTypes();
 
   private static final DeferredRegister<Block> BLOCKS =
       DeferredRegister.create(ForgeRegistries.BLOCKS, Constants.MOD_ID);
@@ -101,7 +100,7 @@ public final class ForgeEasyModelBlockEntityTypes implements EasyModelHostBlockE
                       EasyModelRandomlyAnimatedBlockEntity::new,
                       Set.of(ANIMATED_RANDOMLY_BLOCK.get())));
 
-  private ForgeEasyModelBlockEntityTypes() {}
+  private EasyModelBlockEntityTypes() {}
 
   public static void register(BusGroup modBusGroup) {
     BLOCKS.register(modBusGroup);