+++ Forge/src/main/java/de/markusbordihn/easymodelentities/EasyModelEntitiesClient.java
@@ -21,7 +21,9 @@ package de.markusbordihn.easymodelentities;
 
 import de.markusbordihn.easymodelentities.client.render.EasyModelHostBlockEntityRenderer;
 import de.markusbordihn.easymodelentities.client.render.EasyModelHostEntityRenderer;
+import de.markusbordihn.easymodelentities.client.render.EasyModelSpawnItemSpecialRendererUnbaked;
 import de.markusbordihn.easymodelentities.renderprofile.ModelRenderProfileReloadListener;
+import net.minecraft.client.renderer.special.SpecialModelRenderers;
 import net.minecraftforge.client.event.EntityRenderersEvent;
 import net.minecraftforge.client.event.RegisterClientReloadListenersEvent;
 import org.apache.logging.log4j.LogManager;
@@ -35,29 +37,31 @@ public class EasyModelEntitiesClient {
     log.info("Initializing {} (Forge Client) ...", Constants.MOD_NAME);
     EntityRenderersEvent.RegisterRenderers.BUS.addListener(this::registerEntityRenderers);
     RegisterClientReloadListenersEvent.BUS.addListener(this::registerClientReloadListeners);
+    SpecialModelRenderers.ID_MAPPER.put(
+        EasyModelSpawnItemSpecialRendererUnbaked.ID,
+        EasyModelSpawnItemSpecialRendererUnbaked.MAP_CODEC);
   }
 
   public void registerEntityRenderers(EntityRenderersEvent.RegisterRenderers event) {
     event.registerEntityRenderer(
-        ForgeEasyModelEntityTypes.INSTANCE.groundEntityType(), EasyModelHostEntityRenderer::new);
+        EasyModelEntityTypes.INSTANCE.groundEntityType(), EasyModelHostEntityRenderer::new);
     event.registerEntityRenderer(
-        ForgeEasyModelEntityTypes.INSTANCE.staticEntityType(), EasyModelHostEntityRenderer::new);
+        EasyModelEntityTypes.INSTANCE.staticEntityType(), EasyModelHostEntityRenderer::new);
     event.registerEntityRenderer(
-        ForgeEasyModelEntityTypes.INSTANCE.aquaticEntityType(), EasyModelHostEntityRenderer::new);
+        EasyModelEntityTypes.INSTANCE.aquaticEntityType(), EasyModelHostEntityRenderer::new);
     event.registerEntityRenderer(
-        ForgeEasyModelEntityTypes.INSTANCE.amphibiousEntityType(),
-        EasyModelHostEntityRenderer::new);
+        EasyModelEntityTypes.INSTANCE.amphibiousEntityType(), EasyModelHostEntityRenderer::new);
     event.registerBlockEntityRenderer(
-        ForgeEasyModelBlockEntityTypes.INSTANCE.staticBlockEntityType(),
+        EasyModelBlockEntityTypes.INSTANCE.staticBlockEntityType(),
         EasyModelHostBlockEntityRenderer::new);
     event.registerBlockEntityRenderer(
-        ForgeEasyModelBlockEntityTypes.INSTANCE.tickingBlockEntityType(),
+        EasyModelBlockEntityTypes.INSTANCE.tickingBlockEntityType(),
         EasyModelHostBlockEntityRenderer::new);
     event.registerBlockEntityRenderer(
-        ForgeEasyModelBlockEntityTypes.INSTANCE.animatedBlockEntityType(),
+        EasyModelBlockEntityTypes.INSTANCE.animatedBlockEntityType(),
         EasyModelHostBlockEntityRenderer::new);
     event.registerBlockEntityRenderer(
-        ForgeEasyModelBlockEntityTypes.INSTANCE.animatedRandomlyBlockEntityType(),
+        EasyModelBlockEntityTypes.INSTANCE.animatedRandomlyBlockEntityType(),
         EasyModelHostBlockEntityRenderer::new);
   }